Created RecordActivity/RecordReplayActivity and tried to add functionality to the Rematch and Main Menu buttons, but didn't seem to work.
