--- app/src/main/java/edu/csumb/flailsandfriends/activities/RecordActivity.java
@@ -1,40 +1,0 @@
-package edu.csumb.flailsandfriends.activities;
-
-import android.content.Context;
-import android.content.Intent;
-import android.os.Bundle;
-
-import androidx.activity.EdgeToEdge;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.graphics.Insets;
-import androidx.core.view.ViewCompat;
-import androidx.core.view.WindowInsetsCompat;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import edu.csumb.flailsandfriends.R;
-
-public class RecordActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        EdgeToEdge.enable(this);
-        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main), (v, insets) -> {
-            Insets systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars());
-            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom);
-            return insets;
-        });
-    }
-
-    public void recordSave(String actionLine){
-        List<String> list;
-        list = new ArrayList<String>();
-        list.add(actionLine);
-    }
-
-    static Intent recordActivityIntentFactory(Context context) {
-        return new Intent(context, RecordActivity.class);
-    }
-}
