+++ app/src/main/java/edu/csumb/flailsandfriends/activities/RecordReplayActivity.java
@@ -23,4 +23,6 @@ public class RecordReplayActivity extends AppCompatActivity {
             return insets;
         });
     }
+
+    
 }
