+++ app/src/main/res/layout/activity_record_replay.xml
@@ -7,4 +7,6 @@
     android:layout_height="match_parent"
     tools:context=".activities.RecordReplayActivity">
 
+
+
 </androidx.constraintlayout.widget.ConstraintLayout>
